requirements.txt
ultralytics~=8.0.171
torch~=2.0.1+cu117
numpy~=1.24.1
Pillow~=10.0.1
torchvision~=0.15.2+cu117
flask~=3.0.0
streamlit~=1.27.2